add box drawing and remove junk

+++ src/cp437/converter.rs
@@ -1,4 +1,4 @@
-pub type FontCharType = u16;
+pub type FontCharType = usize;
 
 /// Converts a unicode character to a CP437 equivalent, returning 0 if it didn't have a match
 pub fn to_cp437(c: char) -> FontCharType {
